[#251] refactor class names to increase code understanding
