+++ src/ml_service/feature_extraction/post_processing/post_processing_driver.py
@@ -1,6 +1,6 @@
 from feature_extraction.post_processing.precedent_vector.precedent_vector import PrecedentVector
 from util.file import Load, Save
-from feature_extraction.post_processing.regex import regex_fact
+from feature_extraction.post_processing.regex import regex_entity
 
 def run(command_list):
     """
@@ -8,7 +8,7 @@ def run(command_list):
     :param command_list: Command line arguments
     :return: None
     """
-    regex_fact.run()
+    regex_entity.run()
     fact_cluster_model = Load.load_binary("facts_cluster_model.bin")
     decision_cluster_model = Load.load_binary("decisions_cluster_model.bin")
     fact_data_tuple = Load.load_binary("facts_pre_processed.bin")
